throw rpc exception when exceeding an object limit
Instead of returning an error code. This is something the client
probably isn't expecting to deal with so just blow them up when it
happens. Especially since the limits are just there right now to guard
against buggy scripts. They can try to catch the exception if they
really want to handle it.